requirements.txt
pandas
streamlit
crewai
openai
python-dotenv
